Extract package tests into modules
